:truck: renamed shopify to indexer
