+++ utils/indexer.ts
@@ -1,5 +1,5 @@
 import { ask } from 'https://esm.sh/@kodadot1/uniquery@0.2.0-rc.3'
-// import { getClient } from 'https://esm.sh/@kodadot1/uniquery@0.1.1-rc.0'
+// import { INDEXERS } from "https://esm.sh/@kodadot1/static@0.0.1-rc.0"
 
 const ADDRESS = Deno.env.get("ADDRESS");
 const CHAIN = Deno.env.get("CHAIN");
